requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
annotated-types=0.7.0=pypi_0
blas=1.0=mkl
bottleneck=1.4.2=py310ha9d4c09_0
brotli-python=1.0.9=py310h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.11.26=h06a4308_0
certifi=2024.12.14=py310h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.8=pypi_0
cloudpickle=3.1.0=pypi_0
contourpy=1.3.1=pypi_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cuda-version=12.4=hbda6634_3
cycler=0.12.1=pypi_0
docker-pycreds=0.4.0=pypi_0
farama-notifications=0.0.4=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py310h06a4308_0
fonttools=4.56.0=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.12.0=pypi_0
giflib=5.2.2=h5eee18b_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
grpcio=1.68.1=pypi_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
gymnasium=1.0.0=pypi_0
harl=1.0.0=dev_0
idna=3.7=py310h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.1.4=py310h06a4308_1
jpeg=9e=h5eee18b_3
jsbsim=1.2.1=pypi_0
kiwisolver=1.4.8=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.16=hb9589c4_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libabseil=20240116.2=cxx17_h6a678d5_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.9.1.3=h99ab3db_1
libcurand=10.3.5.147=h99ab3db_1
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.22=h5eee18b_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libopenblas=0.3.21=h043d6bf_0
libpng=1.6.39=h5eee18b_0
libprotobuf=4.25.3=he621ea3_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=hffd6297_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_1
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.7=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.10.1=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.3=py310h06a4308_0
numexpr=2.8.4=py310h8879344_0
numpy=1.24.3=py310hd5efca6_0
numpy-base=1.24.3=py310h8e6c178_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=1.1.1w=h7f8727e_0
packaging=24.2=py310h06a4308_0
pandas=2.2.3=py310h6a678d5_0
pillow=11.0.0=py310hcea889d_1
pip=24.2=py310h06a4308_0
platformdirs=4.3.6=pypi_0
protobuf=5.29.2=pypi_0
psutil=7.0.0=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pygame=2.6.1=pypi_0
pymap3d=3.1.0=pypi_0
pyparsing=3.2.1=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.0=h12debd9_5
python-dateutil=2.9.0post0=py310h06a4308_2
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.5.1=py3.10_cuda11.8_cudnn9.1.0_0
pytorch-cuda=11.8=h7e8668a_6
pytorch-mutex=1.0=cuda
pytz=2024.1=py310h06a4308_0
pyyaml=6.0.2=py310h5eee18b_0
readline=8.2=h5eee18b_0
requests=2.32.3=py310h06a4308_1
scipy=1.14.1=pypi_0
sentry-sdk=2.22.0=pypi_0
setproctitle=1.3.5=pypi_0
setuptools=75.1.0=py310h06a4308_0
six=1.17.0=pypi_0
smmap=5.0.2=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.1=pypi_0
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboardx=2.6.2.2=pypi_0
tk=8.6.14=h39e8969_0
torchaudio=2.5.1=py310_cu118
torchdiffeq=0.2.5=pypi_0
torchtriton=3.1.0=py310
torchvision=0.20.1=py310_cu118
typing_extensions=4.12.2=py310h06a4308_0
tzdata=2024b=h04d1e81_0
urllib3=2.2.3=py310h06a4308_0
wandb=0.19.7=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.44.0=py310h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
